+++ src/main/scala/edu/ie3/simona/agent/grid/GridEnvironment.scala
@@ -6,6 +6,7 @@
 
 package edu.ie3.simona.agent.grid
 
+import edu.ie3.simona.agent.grid.data.GridAgentData.GridAgentRef
 import edu.ie3.simona.agent.participant.ParticipantAgent
 import edu.ie3.simona.model.grid.GridModel
 import org.apache.pekko.actor.typed.ActorRef
@@ -17,25 +18,26 @@ import java.util.UUID
   *
   * @param gridModel
   *   [[GridModel]] with all asset information.
-  * @param nodeToAssetAgents
-  *   A mapping of all node uuids to a set of asset [[ActorRef]] s at those
-  *   nodes.
   * @param inferiorConnections
   *   A map of actor refs to all inferior grids.
   * @param superiorConnections
   *   A map of actor refs to all superior grids with the corresponding superior
   *   nodes.
+  * @param nodeToAssetAgents
+  *   A mapping of all node uuids to a set of asset [[ActorRef]] s at those
+  *   nodes.
+  * @param refToSubgrid
+  *   A mapping of all known references to their subgrid id.
   * @param superiorGridIds
   *   A map of all superior grid uuids to their grid ids.
   */
 final case class GridEnvironment(
     gridModel: GridModel,
+    inferiorConnections: Map[GridAgentRef, Set[UUID]],
+    superiorConnections: Map[GridAgentRef, Set[UUID]],
     nodeToAssetAgents: Map[UUID, Set[ActorRef[ParticipantAgent.Request]]],
-    inferiorConnections: Map[ActorRef[GridAgent.Message], Set[UUID]] =
-      Map.empty,
-    superiorConnections: Map[ActorRef[GridAgent.Message], Set[UUID]] =
-      Map.empty,
-    superiorGridIds: Map[UUID, Int] = Map.empty,
+    refToSubgrid: Map[GridAgentRef, Int],
+    superiorGridIds: Map[UUID, Int],
 ) {
 
   def inferiorNodeUuids: Set[UUID] = inferiorConnections.values.flatten.toSet