+++ src/main/scala/edu/ie3/simona/agent/participant/ParticipantAgentInit.scala
@@ -7,7 +7,6 @@
 package edu.ie3.simona.agent.participant
 
 import edu.ie3.datamodel.models.input.system.SystemParticipantInput
-import edu.ie3.simona.agent.grid.GridAgent
 import edu.ie3.simona.agent.participant.ParticipantAgent.*
 import edu.ie3.simona.agent.{DataInputHandler, SecondaryServiceRegistration}
 import edu.ie3.simona.config.RuntimeConfig.BaseRuntimeConfig
@@ -58,8 +57,6 @@ object ParticipantAgentInit
 
   /** Container class that gathers references to relevant actors.
     *
-    * @param gridAgent
-    *   Reference to the grid agent.
     * @param primaryServiceProxy
     *   Reference to the primary service proxy.
     * @param resultServiceProxy
@@ -68,7 +65,6 @@ object ParticipantAgentInit
     *   References to services by service type.
     */
   final case class ParticipantRefs(
-      gridAgent: ActorRef[GridAgent.Message],
       primaryServiceProxy: ActorRef[ServiceMessage],
       resultServiceProxy: ActorRef[ResultEvent | ExpectResult],
       services: Map[ServiceType, ActorRef[ServiceMessage]],
@@ -296,7 +292,6 @@ object ParticipantAgentInit
       modelShell,
       inputHandler,
       ParticipantGridAdapter(
-        participantRefs.gridAgent,
         simulationParams.expectedPowerRequestTick,
         simulationParams.requestVoltageDeviationTolerance,
       ),