+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentIT.scala
@@ -7,7 +7,6 @@
 package edu.ie3.simona.agent.em
 
 import edu.ie3.datamodel.models.result.system.EmResult
-import edu.ie3.simona.agent.grid.GridAgent
 import edu.ie3.simona.agent.participant.ParticipantAgentInit
 import edu.ie3.simona.agent.participant.ParticipantAgentInit.{
   ParticipantRefs,
@@ -97,7 +96,6 @@ class EmAgentIT
   "An em agent" when {
     "having load, pv and storage agents connected" should {
       "be initialized correctly and run through some activations" in {
-        val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultServiceProxy =
           TestProbe[ResultEvent | ExpectResult]("ResultServiceProxy")
         val primaryServiceProxy =
@@ -106,7 +104,6 @@ class EmAgentIT
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
-          gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryServiceProxy.ref,
           resultServiceProxy = resultServiceProxy.ref,
           services = Map(ServiceType.WeatherService -> weatherService.ref),
@@ -384,7 +381,6 @@ class EmAgentIT
 
     "having load, pv and heat pump agents connected" should {
       "be initialized correctly and run through some activations" in {
-        val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultServiceProxy =
           TestProbe[ResultEvent | ExpectResult]("ResultServiceProxy")
         val primaryServiceProxy =
@@ -393,7 +389,6 @@ class EmAgentIT
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
-          gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryServiceProxy.ref,
           resultServiceProxy = resultServiceProxy.ref,
           services = Map(ServiceType.WeatherService -> weatherService.ref),
@@ -952,7 +947,6 @@ class EmAgentIT
 
     "having a pv and a load agent connected" should {
       "have correct values also for agents with limited operation time" in {
-        val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultServiceProxy =
           TestProbe[ResultEvent | ExpectResult]("ResultServiceProxy")
         val primaryServiceProxy =
@@ -961,7 +955,6 @@ class EmAgentIT
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
-          gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryServiceProxy.ref,
           resultServiceProxy = resultServiceProxy.ref,
           services = Map(ServiceType.WeatherService -> weatherService.ref),