+++ CHANGELOG.md
@@ -79,6 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cleaned up EVCS test data [#1690](https://github.com/ie3-institute/simona/issues/1690)
 - Simplified Option handling in PowerFlowSupport [#1701](https://github.com/ie3-institute/simona/issues/1701)
 - Removed lowest SOC, added departure SOC target for EVs [#1096](https://github.com/ie3-institute/simona/issues/1096)
+- Refactor initialization of `GridAgent` [#1686](https://github.com/ie3-institute/simona/issues/1686)
 
 ### Fixed
 - Fixes in Documentation, ScalaDocs, Code Style and more [#1397](https://github.com/ie3-institute/simona/issues/1397)